@@ -1,28 +1,24 @@
 <Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>1.0.3</Version>
-    <Company>cl2j</Company>
-    <Product>cl2j.FileStorage</Product>
-    <Description>cl2j.FileStorage is a multi-providers library that abstract file operations like read, write, delete and more. It's an open and extensible framework based on interfaces and Dependency Injection.
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+        <Version>1.0.4</Version>
+        <Company>cl2j</Company>
+        <Product>cl2j.FileStorage</Product>
+        <Description>cl2j.FileStorage is a multi-providers library that abstract file operations like read, write, delete and more. It's an open and extensible framework based on interfaces and Dependency Injection.
 
 Azure Blob Storage implementation</Description>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/jfblier/cl2j.FileStorage</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/jfblier/cl2j.FileStorage</RepositoryUrl>
-    <RepositoryType>GitHub</RepositoryType>
-    <PackageReadmeFile>readme.md</PackageReadmeFile>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Azure.Storage.Blobs" Version="12.10.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\cl2j.FileStorage\cl2j.FileStorage.csproj" />
-
-    <None Include="..\readme.md" Pack="true" PackagePath="\" />
-  </ItemGroup>
+        <PackageLicenseExpression>MIT</PackageLicenseExpression>
+        <PackageProjectUrl>https://github.com/jfblier/cl2j.FileStorage</PackageProjectUrl>
+        <RepositoryUrl>https://github.com/jfblier/cl2j.FileStorage</RepositoryUrl>
+        <RepositoryType>GitHub</RepositoryType>
+        <PackageReadmeFile>readme.md</PackageReadmeFile>
+    </PropertyGroup>
+    <ItemGroup>
+        <PackageReference Include="Azure.Storage.Blobs" Version="12.10.0" />
+    </ItemGroup>
+    <ItemGroup>
+        <ProjectReference Include="..\cl2j.FileStorage\cl2j.FileStorage.csproj" />
+        <None Include="..\readme.md" Pack="true" PackagePath="\" />
+    </ItemGroup>
 </Project>
