@@ -1,20 +1,17 @@
 ﻿using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
-using System.Collections.Generic;
 
 namespace cl2j.FileStorage.Core
 {
     internal class FileStorageFactory : IFileStorageFactory
     {
         private static readonly object Lock = new();
+        public List<string> registrations = new();
         private readonly IConfigurationRoot configuration;
-        private readonly ILogger<FileStorageFactory> logger;
-        private static Dictionary<string, IFileStorageProvider> storageProviderInstances = new Dictionary<string, IFileStorageProvider>();
+        private static readonly Dictionary<string, IFileStorageProvider> storageProviderInstances = new();
 
-        public FileStorageFactory(IConfigurationRoot configuration, ILogger<FileStorageFactory> logger)
+        public FileStorageFactory(IConfigurationRoot configuration)
         {
             this.configuration = configuration;
-            this.logger = logger;
         }
 
         public void Register<T>(string type) where T : IFileStorageProvider, new()
@@ -35,13 +32,15 @@ namespace cl2j.FileStorage.Core
 
                         storageProviderInstances.Add(providerName, provider);
 
-                        logger?.LogTrace($"Registered '{providerName}' with provider '{typeof(T).Name}'");
+                        registrations.Add($"'{providerName}' with provider '{typeof(T).Name}'");
                     }
                 }
             }
         }
 
-        public IFileStorageProvider Get(string name)
+        public IList<string> Registrations => registrations;
+
+        public IFileStorageProvider? Get(string name)
         {
             //First search into the cached fileSystems
             if (storageProviderInstances.TryGetValue(name, out var fileStorageInstance))
@@ -52,20 +51,8 @@ namespace cl2j.FileStorage.Core
             {
                 if (storageProviderInstances.TryGetValue(name, out fileStorageInstance))
                     return fileStorageInstance;
-
-                //if (configurations.TryGetValue(name, out var configuration))
-                //{
-                //    fileSystem = Create(configuration);
-                //    stores.Add(name, fileSystem);
-                //    return fileSystem;
-                //}
             }
 
-            //var knownFileSystems = new StringBuilder();
-            //foreach (var kvp in configurations)
-            //    knownFileSystems.Append($"'{kvp.Key}' ");
-
-            //throw new Exception($"FileSystem '{name}' is not defined in the application bootstrap. Known file systems: {knownFileSystems}");
             return null;
         }
     }
