@@ -2,7 +2,7 @@
 {
     public class FileStorageConfiguration
     {
-        public string Type { get; set; }
-        public string Name { get; set; }
+        public string Type { get; set; } = null!;
+        public string Name { get; set; } = null!;
     }
 }
