@@ -1,12 +1,20 @@
 ﻿using cl2j.FileStorage.Core;
 using System.Text;
 using System.Text.Json;
-using System.Threading.Tasks;
+using System.Text.Json.Serialization;
 
 namespace cl2j.FileStorage.Extensions
 {
     public static class JsonExtensions
     {
+        private static readonly JsonSerializerOptions optionsDeserialize = new();
+
+        static JsonExtensions()
+        {
+            optionsDeserialize.PropertyNameCaseInsensitive = true;
+            optionsDeserialize.Converters.Add(new JsonStringEnumConverter());
+        }
+
         public static string SerializeJsonObject<T>(T obj, bool indented = false)
         {
             var options = new JsonSerializerOptions()
@@ -17,16 +25,12 @@ namespace cl2j.FileStorage.Extensions
             return JsonSerializer.Serialize(obj, options);
         }
 
-        public static T DeserializeJsonObject<T>(string json)
+        public static T? DeserializeJsonObject<T>(string json)
         {
-            var options = new JsonSerializerOptions()
-            {
-                PropertyNameCaseInsensitive = true
-            };
-            return JsonSerializer.Deserialize<T>(json, options);
+            return JsonSerializer.Deserialize<T>(json, optionsDeserialize);
         }
 
-        public static async Task<T> ReadJsonObjectAsync<T>(this IFileStorageProvider fileStorageProvider, string fileName, Encoding encoding = null) where T : new()
+        public static async Task<T?> ReadJsonObjectAsync<T>(this IFileStorageProvider fileStorageProvider, string fileName, Encoding? encoding = null) where T : new()
         {
             var data = await fileStorageProvider.ReadTextAsync(fileName, encoding);
 
@@ -36,7 +40,7 @@ namespace cl2j.FileStorage.Extensions
             return DeserializeJsonObject<T>(data);
         }
 
-        public static async Task WriteJsonObjectAsync<T>(this IFileStorageProvider fileStorageProvider, string fileName, T obj, bool indented = false, Encoding encoding = null)
+        public static async Task WriteJsonObjectAsync<T>(this IFileStorageProvider fileStorageProvider, string fileName, T obj, bool indented = false, Encoding? encoding = null)
         {
             var data = SerializeJsonObject(obj, indented);
             await fileStorageProvider.WriteTextAsync(fileName, data, encoding);
