@@ -4,6 +4,6 @@ namespace cl2j.FileStorage.Provider.Disk
 {
     public class FileStorageProviderDiskConfiguration : FileStorageConfiguration
     {
-        public string Path { get; set; }
+        public string Path { get; set; } = null!;
     }
 }
